[#1022] Resolve the named account in one place across the administrative endpoints
Eight administrative endpoints asked the same two questions of the account a
request named — is there one, and is it one this deployment serves — and
answered a caller who got it wrong in their own copy of the same sentence. Six
files carried an identical ten-line ResolveAccount, five an UnknownAccount with
two blank-account variants, and two inlined both checks. The copies had already
drifted: the two inlined ones echoed the normalized identifier while the other
six echoed the request text, so a whitespace-padded name was answered
differently depending on which route it was sent to.

AdminAccountRequest now holds the guard: Resolve reads the account, Refuse picks
between Missing and Unknown, MissingFilter states the optional-filter case, and
TryResolveFilter carries the shape two readings use, where an absent filter is
every account rather than a refusal. The whitespace question is settled toward
the identifier that was looked up, so the refusal names what MailFathom searched
for and no request text reaches the response body unaltered.

The response and request records of OutboxEndpoints.cs move to an
OutboxResponses.cs companion, which is the convention five endpoint families
already follow and leaves the 540-line file as the routes and the handlers.

No route, verb, status code, problem type, permission, request shape or response
shape changes.
